fix: ajuste provider
