+++ react/src/components/store-provider/index.tsx
@@ -1,28 +1,7 @@
 import * as React from 'react'
-import { Provider, useDispatch } from 'react-redux'
-import { socket } from '@tray-storefront/core'
+import { Provider } from 'react-redux'
 import store from '../../store/configure-store'
-import { updateSettings } from '../../store/modules/settings'
-import { updateSections } from '../../store/modules/sections'
-
-function RootProvider(props) {
-  const urlParams = new URLSearchParams(window.location.search)
-  const hashPreview = urlParams.get('preview')
-  const dispatch = useDispatch()
-
-  function onUpdate({ shopID, data }) {
-    if (data) {
-      dispatch(updateSettings(data?.settings))
-      dispatch(updateSections(data?.sections))
-    }
-  }
-
-  if (hashPreview) {
-    socket.create(hashPreview, onUpdate)
-  }
-
-  return <>{props.children}</>
-}
+import RootProvider from './root-provider'
 
 export function StoreProvider(props) {
   return (
