+++ react/src/hooks/use-settings.tsx
@@ -13,5 +13,5 @@ export function useSettings() {
     }
   }, [dispatch])
 
-  return { settings }
+  return settings
 }
